Revert "Move to bytes.Buffer"
This reverts commit 82523653d07a14300feaa7da71730f855523f9c4.

+++ utils.go
@@ -2,18 +2,18 @@ package main
 
 import "bytes"
 
-func indexMultiple(buf []byte, patterns ...[][]byte) [][][2]int64 {
-	result := [][][2]int64{}
+func indexMultiple(buf []byte, patterns ...[][]byte) [][][2]int {
+	result := [][][2]int{}
 	bufPos := 0
 
 	for _, pattern := range patterns {
-		ptnResult := [][2]int64{}
+		ptnResult := [][2]int{}
 		for _, ptn := range pattern {
 			pos := bytes.Index(buf[bufPos:], ptn)
 			if pos == -1 {
 				return result
 			}
-			ptnResult = append(ptnResult, [2]int64{int64(bufPos + pos), int64(bufPos + pos + len(ptn))})
+			ptnResult = append(ptnResult, [2]int{bufPos + pos, bufPos + pos + len(ptn)})
 			bufPos += pos + len(ptn)
 		}
 		if len(ptnResult) == 0 {
